changed name of java target package from default to EventDrivenJavaTargetPackage
