+++ Java Target/EventBasedJavaTarget/src/EventDrivenJavaTargetPackage/Driver.java
@@ -1,3 +1,4 @@
+package EventDrivenJavaTargetPackage;
 import java.util.Arrays;
 import java.util.Scanner;
